+++ src/react/navigation/index.ts
@@ -1,4 +1,4 @@
 export { cellNavigation } from './addon';
 export { CELL_NAVIGATION_ADDON, cellNavigationMessages } from './messages';
-export { useCellNavigation, useOptionalCellNavigation } from './context';
+export { useCellNavigation, useCellTabIndex, useOptionalCellNavigation } from './context';
 export type { ActiveCell, CellNavigationOptions, CellNavigationController } from './types';